+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.8] - 2021-05-26
+### Bug Fixes
+#### com.unity.ml-agents (C#)
+- Fixed a null reference exception that occurred when loading an ONNX model file that was generated with a new
+version of the Python trainer (0.26.0 or newer). (#5350)
+- Added checked to prevent training with incompatible versions of the python trainer (0.26.0 or newer). (#5370)
+
 ## [1.0.7] - 2021-03-04
 ### Minor Changes
 #### com.unity.ml-agents (C#)